requirements.txt
Django~=4.2.4
djangorestframework~=3.14.0
psycopg2-binary==2.9.1
djangorestframework-jwt==1.11.0
django-cors-headers==3.10.0
